requirements.txt
python_vlc==3.0.102
